server/requirements.txt
psycopg2
paho-mqtt
pycryptodome
marshmallow
pandas
pystan
fbprophet
requests


